fix(dataset): create new dataset
