+++ data_release/ground_truth.csv
@@ -1,0 +1,101 @@
+Image Name,MOS,AesA3_Score,AesA2_Score,AesA1_Score
+ss01 (847).jpg,0.1776,2.5984,1.7104,Low
+ss01 (1772).jpg,0.223,3.007,1.892,Low
+ss02 (512).jpg,0.2194,2.9746,1.8776,Low
+ss01 (151).jpg,0.1931,2.7379,1.7724,Low
+ss03 (7636).jpg,0.1755,2.5795,1.702,Low
+ss01 (1214).jpg,0.1942,2.7478,1.7768,Low
+ss01 (632).jpg,0.2922,3.6298,2.1688,Low
+ss02 (755).jpg,0.2822,3.5398,2.1288,Low
+ss01 (178).jpg,0.275,3.475,2.1,Low
+ss01 (285).jpg,0.245,3.205,1.98,Low
+ss01 (559).jpg,0.2295,3.0655,1.918,Low
+ss01 (1067).jpg,0.2462,3.2158,1.9848,Low
+ss03 (6337).jpg,0.3179,3.8611,2.2716,Low
+ss03 (6315).jpg,0.2797,3.5173,2.1188,Low
+ss02 (338).jpg,0.3032,3.7288,2.2128,Low
+ss01 (565).jpg,0.3162,3.8458,2.2648,Low
+ss01 (1714).jpg,0.2168,2.9512,1.8672,Low
+ss01 (333).jpg,0.2604,3.3436,2.0416,Low
+ss02 (1649).jpg,0.3033,3.7297,2.2132,Low
+ss01 (1709).jpg,0.2585,3.3265,2.034,Low
+ss01 (450).jpg,0.198,2.782,1.792,Low
+ss01 (1125).jpg,0.2275,3.0475,1.91,Low
+ss01 (1465).jpg,0.3199,3.8791,2.2796,Low
+ss02 (930).jpg,0.2314,3.0826,1.9256,Low
+ss02 (1761).jpg,0.1977,2.7793,1.7908,Low
+ss01 (307).jpg,0.2308,3.0772,1.9232,Low
+ss01 (1869).jpg,0.2989,3.6901,2.1956,Low
+ss01 (624).jpg,0.2807,3.5263,2.1228,Low
+ss01 (687).jpg,0.2204,2.9836,1.8816,Low
+ss02 (617).jpg,0.2394,3.1546,1.9576,Low
+ss01 (464).jpg,0.2651,3.3859,2.0604,Low
+ss01 (1917).jpg,0.2487,3.2383,1.9948,Low
+ss01 (1493).jpg,0.2377,3.1393,1.9508,Low
+ss02 (1083).jpg,0.2948,3.6532,2.1792,Low
+ss03 (448).jpg,0.5436,5.8924,3.1744,Medium
+ss03 (6531).jpg,0.4262,4.8358,2.7048,Medium
+ss03 (7942).jpg,0.4933,5.4397,2.9732,Medium
+ss02 (144).jpg,0.6083,6.4747,3.4332,Medium
+ss03 (7681).jpg,0.5766,6.1894,3.3064,Medium
+ss03 (5601).jpg,0.5895,6.3055,3.358,Medium
+ss03 (5954).jpg,0.5078,5.5702,3.0312,Medium
+ss02 (411).jpg,0.4861,5.3749,2.9444,Medium
+ss03 (6462).jpg,0.4272,4.8448,2.7088,Medium
+ss02 (562).jpg,0.358,4.222,2.432,Medium
+ss01 (13).jpg,0.6421,6.7789,3.5684,Medium
+ss02 (696).jpg,0.5815,6.2335,3.326,Medium
+ss02 (299).jpg,0.4328,4.8952,2.7312,Medium
+ss02 (189).jpg,0.4678,5.2102,2.8712,Medium
+ss01 (258).jpg,0.5829,6.2461,3.3316,Medium
+ss03 (5851).jpg,0.6128,6.5152,3.4512,Medium
+ss01 (1992).jpg,0.508,5.572,3.032,Medium
+ss03 (1993).jpg,0.618,6.562,3.472,Medium
+ss02 (210).jpg,0.4237,4.8133,2.6948,Medium
+ss01 (890).jpg,0.4903,5.4127,2.9612,Medium
+ss01 (1799).jpg,0.3617,4.2553,2.4468,Medium
+ss02 (202).jpg,0.4054,4.6486,2.6216,Medium
+ss03 (7628).jpg,0.6313,6.6817,3.5252,Medium
+ss02 (1214).jpg,0.4869,5.3821,2.9476,Medium
+ss02 (762).jpg,0.5832,6.2488,3.3328,Medium
+ss03 (2000).jpg,0.632,6.688,3.528,Medium
+ss03 (5713).jpg,0.6341,6.7069,3.5364,Medium
+ss01 (1187).jpg,0.4189,4.7701,2.6756,Medium
+ss03 (925).jpg,0.3991,4.5919,2.5964,Medium
+ss02 (910).jpg,0.4323,4.8907,2.7292,Medium
+ss01 (1470).jpg,0.4338,4.9042,2.7352,Medium
+ss01 (112).jpg,0.6407,6.7663,3.5628,Medium
+ss01 (1017).jpg,0.4412,4.9708,2.7648,Medium
+ss03 (2459).jpg,0.8217,8.3953,4.2868,High
+ss03 (4350).jpg,0.8515,8.6635,4.406,High
+ss03 (7124).jpg,0.8717,8.8453,4.4868,High
+ss03 (4675).jpg,0.8027,8.2243,4.2108,High
+ss03 (51).jpg,0.7927,8.1343,4.1708,High
+ss03 (5093).jpg,0.8386,8.5474,4.3544,High
+ss03 (1104).jpg,0.7767,7.9903,4.1068,High
+ss03 (4818).jpg,0.8076,8.2684,4.2304,High
+ss03 (5367).jpg,0.8214,8.3926,4.2856,High
+ss01 (1397).jpg,0.6792,7.1128,3.7168,High
+ss03 (4609).jpg,0.8164,8.3476,4.2656,High
+ss03 (3275).jpg,0.6842,7.1578,3.7368,High
+ss03 (7417).jpg,0.7594,7.8346,4.0376,High
+ss03 (1387).jpg,0.71,7.39,3.84,High
+ss03 (7128).jpg,0.8498,8.6482,4.3992,High
+ss03 (1090).jpg,0.8903,9.0127,4.5612,High
+ss03 (2663).jpg,0.7172,7.4548,3.8688,High
+ss03 (4255).jpg,0.7874,8.0866,4.1496,High
+ss03 (6992).jpg,0.7881,8.0929,4.1524,High
+ss03 (7674).jpg,0.749,7.741,3.996,High
+ss03 (4791).jpg,0.8096,8.2864,4.2384,High
+ss03 (2564).jpg,0.7733,7.9597,4.0932,High
+ss03 (4045).jpg,0.8569,8.7121,4.4276,High
+ss03 (7790).jpg,0.8226,8.4034,4.2904,High
+ss03 (695).jpg,0.758,7.822,4.032,High
+ss03 (4630).jpg,0.7502,7.7518,4.0008,High
+ss03 (2955).jpg,0.7538,7.7842,4.0152,High
+ss03 (5177).jpg,0.8024,8.2216,4.2096,High
+ss03 (4538).jpg,0.8802,8.9218,4.5208,High
+ss03 (6448).jpg,0.7376,7.6384,3.9504,High
+ss03 (742).jpg,0.7522,7.7698,4.0088,High
+ss03 (4442).jpg,0.8397,8.5573,4.3588,High
+ss03 (3640).jpg,0.7768,7.9912,4.1072,High
