+++ data_release/mos_ground_truth.csv
@@ -1,0 +1,101 @@
+Image Name,MOS
+ss01 (847).jpg,0.1776
+ss01 (1772).jpg,0.223
+ss02 (512).jpg,0.2194
+ss01 (151).jpg,0.1931
+ss03 (7636).jpg,0.1755
+ss01 (1214).jpg,0.1942
+ss01 (632).jpg,0.2922
+ss02 (755).jpg,0.2822
+ss01 (178).jpg,0.275
+ss01 (285).jpg,0.245
+ss01 (559).jpg,0.2295
+ss01 (1067).jpg,0.2462
+ss03 (6337).jpg,0.3179
+ss03 (6315).jpg,0.2797
+ss02 (338).jpg,0.3032
+ss01 (565).jpg,0.3162
+ss01 (1714).jpg,0.2168
+ss01 (333).jpg,0.2604
+ss02 (1649).jpg,0.3033
+ss01 (1709).jpg,0.2585
+ss01 (450).jpg,0.198
+ss01 (1125).jpg,0.2275
+ss01 (1465).jpg,0.3199
+ss02 (930).jpg,0.2314
+ss02 (1761).jpg,0.1977
+ss01 (307).jpg,0.2308
+ss01 (1869).jpg,0.2989
+ss01 (624).jpg,0.2807
+ss01 (687).jpg,0.2204
+ss02 (617).jpg,0.2394
+ss01 (464).jpg,0.2651
+ss01 (1917).jpg,0.2487
+ss01 (1493).jpg,0.2377
+ss02 (1083).jpg,0.2948
+ss03 (448).jpg,0.5436
+ss03 (6531).jpg,0.4262
+ss03 (7942).jpg,0.4933
+ss02 (144).jpg,0.6083
+ss03 (7681).jpg,0.5766
+ss03 (5601).jpg,0.5895
+ss03 (5954).jpg,0.5078
+ss02 (411).jpg,0.4861
+ss03 (6462).jpg,0.4272
+ss02 (562).jpg,0.358
+ss01 (13).jpg,0.6421
+ss02 (696).jpg,0.5815
+ss02 (299).jpg,0.4328
+ss02 (189).jpg,0.4678
+ss01 (258).jpg,0.5829
+ss03 (5851).jpg,0.6128
+ss01 (1992).jpg,0.508
+ss03 (1993).jpg,0.618
+ss02 (210).jpg,0.4237
+ss01 (890).jpg,0.4903
+ss01 (1799).jpg,0.3617
+ss02 (202).jpg,0.4054
+ss03 (7628).jpg,0.6313
+ss02 (1214).jpg,0.4869
+ss02 (762).jpg,0.5832
+ss03 (2000).jpg,0.632
+ss03 (5713).jpg,0.6341
+ss01 (1187).jpg,0.4189
+ss03 (925).jpg,0.3991
+ss02 (910).jpg,0.4323
+ss01 (1470).jpg,0.4338
+ss01 (112).jpg,0.6407
+ss01 (1017).jpg,0.4412
+ss03 (2459).jpg,0.8217
+ss03 (4350).jpg,0.8515
+ss03 (7124).jpg,0.8717
+ss03 (4675).jpg,0.8027
+ss03 (51).jpg,0.7927
+ss03 (5093).jpg,0.8386
+ss03 (1104).jpg,0.7767
+ss03 (4818).jpg,0.8076
+ss03 (5367).jpg,0.8214
+ss01 (1397).jpg,0.6792
+ss03 (4609).jpg,0.8164
+ss03 (3275).jpg,0.6842
+ss03 (7417).jpg,0.7594
+ss03 (1387).jpg,0.71
+ss03 (7128).jpg,0.8498
+ss03 (1090).jpg,0.8903
+ss03 (2663).jpg,0.7172
+ss03 (4255).jpg,0.7874
+ss03 (6992).jpg,0.7881
+ss03 (7674).jpg,0.749
+ss03 (4791).jpg,0.8096
+ss03 (2564).jpg,0.7733
+ss03 (4045).jpg,0.8569
+ss03 (7790).jpg,0.8226
+ss03 (695).jpg,0.758
+ss03 (4630).jpg,0.7502
+ss03 (2955).jpg,0.7538
+ss03 (5177).jpg,0.8024
+ss03 (4538).jpg,0.8802
+ss03 (6448).jpg,0.7376
+ss03 (742).jpg,0.7522
+ss03 (4442).jpg,0.8397
+ss03 (3640).jpg,0.7768
