--- data_release/obsolete_ground_truth.csv
@@ -1,101 +1,0 @@
-Image Name,MOS, Level
-ss03 (1547).jpg,0.8507,High
-ss03 (6966).jpg,0.8056,High
-ss03 (7458).jpg,0.8115,High
-ss03 (2798).jpg,0.8159,High
-ss03 (6338).jpg,0.8712,High
-ss03 (68).jpg,0.8244,High
-ss03 (3451).jpg,0.8029,High
-ss03 (7082).jpg,0.8594,High
-ss03 (3061).jpg,0.8141,High
-ss03 (7896).jpg,0.8732,High
-ss03 (2203).jpg,0.8018,High
-ss03 (1892).jpg,0.8276,High
-ss03 (3869).jpg,0.8279,High
-ss03 (5448).jpg,0.813,High
-ss03 (4857).jpg,0.8266,High
-ss03 (5100).jpg,0.8106,High
-ss03 (3008).jpg,0.8042,High
-ss03 (7515).jpg,0.845,High
-ss03 (5547).jpg,0.8215,High
-ss03 (4973).jpg,0.8325,High
-ss03 (1572).jpg,0.8295,High
-ss03 (1320).jpg,0.8574,High
-ss03 (6692).jpg,0.8037,High
-ss03 (5262).jpg,0.8301,High
-ss03 (1244).jpg,0.8449,High
-ss03 (6869).jpg,0.8084,High
-ss03 (6833).jpg,0.8056,High
-ss03 (2930).jpg,0.835,High
-ss03 (4178).jpg,0.8158,High
-ss03 (1486).jpg,0.8159,High
-ss03 (5101).jpg,0.8447,High
-ss03 (4887).jpg,0.8076,High
-ss03 (1045).jpg,0.8259,High
-ss03 (1640).jpg,0.836,High
-ss02 (886).jpg,0.4745,Medium
-ss02 (604).jpg,0.6188,Medium
-ss03 (805).jpg,0.6337,Medium
-ss03 (2904).jpg,0.6463,Medium
-ss03 (5466).jpg,0.6457,Medium
-ss03 (7528).jpg,0.6803,Medium
-ss01 (1286).jpg,0.6263,Medium
-ss03 (6443).jpg,0.6018,Medium
-ss01 (1614).jpg,0.4933,Medium
-ss02 (1814).jpg,0.5128,Medium
-ss03 (6516).jpg,0.6422,Medium
-ss03 (120).jpg,0.5422,Medium
-ss01 (1544).jpg,0.4868,Medium
-ss03 (1548).jpg,0.6918,Medium
-ss03 (557).jpg,0.567,Medium
-ss03 (321).jpg,0.5308,Medium
-ss01 (1002).jpg,0.5124,Medium
-ss03 (3738).jpg,0.6902,Medium
-ss03 (37).jpg,0.6442,Medium
-ss03 (6061).jpg,0.5379,Medium
-ss02 (1109).jpg,0.5983,Medium
-ss03 (4515).jpg,0.6401,Medium
-ss03 (7181).jpg,0.655,Medium
-ss02 (182).jpg,0.5139,Medium
-ss01 (535).jpg,0.6944,Medium
-ss02 (1500).jpg,0.5153,Medium
-ss03 (2661).jpg,0.4767,Medium
-ss03 (604).jpg,0.6399,Medium
-ss03 (6534).jpg,0.6518,Medium
-ss03 (6170).jpg,0.511,Medium
-ss02 (334).jpg,0.5633,Medium
-ss02 (1971).jpg,0.4927,Medium
-ss01 (1737).jpg,0.6458,Medium
-ss02 (1466).jpg,0.3022,Low
-ss02 (1764).jpg,0.386,Low
-ss01 (10).jpg,0.2896,Low
-ss01 (1785).jpg,0.1582,Low
-ss01 (1399).jpg,0.3373,Low
-ss01 (617).jpg,0.1882,Low
-ss01 (544).jpg,0.2731,Low
-ss01 (741).jpg,0.2017,Low
-ss01 (482).jpg,0.1957,Low
-ss02 (602).jpg,0.2869,Low
-ss01 (78).jpg,0.1462,Low
-ss01 (1939).jpg,0.2259,Low
-ss01 (1696).jpg,0.2172,Low
-ss01 (1552).jpg,0.3466,Low
-ss01 (1914).jpg,0.2374,Low
-ss01 (1445).jpg,0.2034,Low
-ss02 (880).jpg,0.3101,Low
-ss01 (1927).jpg,0.1926,Low
-ss03 (5945).jpg,0.3496,Low
-ss01 (1784).jpg,0.1872,Low
-ss01 (427).jpg,0.2571,Low
-ss01 (1701).jpg,0.3696,Low
-ss02 (1506).jpg,0.3233,Low
-ss01 (1093).jpg,0.1618,Low
-ss02 (1763).jpg,0.2095,Low
-ss02 (219).jpg,0.161,Low
-ss01 (1369).jpg,0.1635,Low
-ss02 (1335).jpg,0.2186,Low
-ss01 (1051).jpg,0.2349,Low
-ss02 (141).jpg,0.3553,Low
-ss02 (895).jpg,0.2109,Low
-ss01 (1114).jpg,0.2497,Low
-ss01 (467).jpg,0.3733,Low
